requirements.txt
streamlit
PyMySQL
sqlalchemy
python-dotenv
captcha
streamlit-paste-button
